Fix floating point issues by making time step an integer

+++ experiment_control/experiment_control.compute_device_parameters/experiment_control/compute_device_parameters/clock_instruction.py
@@ -8,7 +8,7 @@ class ClockInstruction:
 
     start: float
     stop: float
-    time_step: float
+    time_step: int
     order: "StepInstruction"
 
     class StepInstruction(Enum):
